working database implementation and floats changed to doubles. not sure if SQLite.swift file is being included

+++ Nutrient+/NutrientPlus/AddFoods.swift
@@ -113,7 +113,7 @@ extension AddFoods: UITableViewDataSource, UITableViewDelegate{
         let numberServings = (self.numberOfServings?.text)!
         //print(mainViewController.nutrients)
         for items in AddFoods.nutrientCards{
-            let currentNutrientAmount = Float(items.amount) * Float(numberServings)!
+            let currentNutrientAmount = Double(items.amount) * Double(numberServings)!
             if (mainViewController.nutrients.keys.contains(items.nutrientName)){
                 mainViewController.nutrients[items.nutrientName] = mainViewController.nutrients[items.nutrientName]! + currentNutrientAmount
             }
